refactor(section): Line-Section 의존성 단방향 되도록 변경
