--- src/main/java/nextstep/subway/domain/SectionRepository.java
@@ -1,9 +1,0 @@
-package nextstep.subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface SectionRepository extends JpaRepository<Section, Long> {
-    List<Section> findByLineId(Long lineId);
-}
