+++ src/main/java/nextstep/subway/ui/LineController.java
@@ -1,7 +1,6 @@
 package nextstep.subway.ui;
 
 import nextstep.subway.application.LineService;
-import nextstep.subway.application.SectionService;
 import nextstep.subway.dto.LineRequest;
 import nextstep.subway.dto.LineResponse;
 import nextstep.subway.dto.SectionRequest;
@@ -17,11 +16,9 @@ import java.util.List;
 @RestController
 public class LineController {
     private final LineService lineService;
-    private final SectionService sectionService;
 
-    public LineController(LineService lineService, SectionService sectionService) {
+    public LineController(LineService lineService) {
         this.lineService = lineService;
-        this.sectionService = sectionService;
     }
 
     @PostMapping("/lines")
@@ -54,13 +51,13 @@ public class LineController {
     @PostMapping(value = "/lines/{lineId}/sections")
     public ResponseEntity<SectionResponse> createSections(@PathVariable final Long lineId,
                                                              @RequestBody SectionRequest sectionRequest) {
-        SectionResponse section = sectionService.saveSection(lineId, sectionRequest);
+        SectionResponse section = lineService.addSection(lineId, sectionRequest);
         return ResponseEntity.created(URI.create("/lines/" + section.getId() + "/sections")).body(section);
     }
 
     @GetMapping(value = "/lines/{lineId}/sections")
     public ResponseEntity<List<SectionResponse>> showSections(@PathVariable final Long lineId) {
-        return ResponseEntity.ok(sectionService.findAllSections(lineId));
+        return ResponseEntity.ok(lineService.findAllSections(lineId));
     }
 
     @ExceptionHandler({DataIntegrityViolationException.class, IllegalArgumentException.class})
